chore: made several revisions from feedback. missing discussion on directories and simplifying lesson 7
