+++ src/static/courses.json
@@ -1,4 +1,4 @@
 {
   "all": ["0004", "0005", "0001", "0002", "0003"],
-  "featured": ["0004", "0001", "0002", "0003"]
+  "featured": ["0004", "0001", "0002", "0003", "0005"]
 }
