+++ src/tutorials/0005/02-exercise.md
@@ -1,4 +1,1 @@
-Let's try to add a file to your IPFS node! First, try to upload one or more files.
-
-**Hint:** You can pass either a `File` or an array of `File`s to the `add` method.
-
+First, upload one or more files by dragging and dropping below or clicking to make a selection from your file explorer.  Next, in the code editor, remove the comment markers (`//`) that precede `return files` within the `run` function.
