+++ src/tutorials/0005/02.md
@@ -1,39 +1,12 @@
+As a matter of security, web browsers don't let us directly change files that live in your computer's file system. Therefore you'll need to upload one or more files to the browser that you can use throughout this tutorial.
 
-## Working with files in ProtoSchool
-Here in our ProtoSchool tutorials, we create a new IPFS node for you in the browser each time you hit the "Submit" button in a lesson. Whenever you see `ipfs.someMethod()` in our lessons, `ipfs` is a variable that refers to your IPFS instance, also known as a node. The actions that you take only affect your own IPFS node, not nodes belonging to your peers.
+Within each exercise, you'll see that you can upload files from your computer either by dragging and dropping or selecting them from your file explorer. If you look closely at the `run` function in the code editor, you'll notice that it now takes an argument `files`. When you upload files from your computer, we'll make sure they're passed into the function as the `files` array. As long as you don't refresh your browser, these files will remain accessible for the next lesson in the tutorial, but you'll also have the option to upload different files to work with for each lesson.
 
-We create this special IPFS node behind the scenes, so you can focus on the content of our lessons. To host a fully functional IPFS node of your own, though, IPFS must be installed on your machine, and a local daemon running in your terminal. If you haven't already experimented with this, visit our docs to learn how to [install IPFS](https://docs.ipfs.io/guides/guides/install/) and [initialize your node](https://docs.ipfs.io/introduction/usage/#initialize-the-repository).
+Each element in this `files` array will be of type `File`. You can find more about the `File` type [here](https://developer.mozilla.org/en-US/docs/Web/API/File). Aside from the contents of the uploaded file, a `File` type also contains the following attributes:
 
-As mentioned previously, the methods discussed in this tutorial are part of the [Files API](https://github.com/ipfs/interface-js-ipfs-core/blob/master/SPEC/FILES.md). Check the documentation for more specific details, such as options for each API function.
+* **name** (name of the uploaded file)
+* **lastModified** (date the uploaded file was last modified in)
+* **size** (size of the uploaded file)
+* **type** (type of the uploaded file)
 
-## Uploading files in the exercise
-
-As we will be working with adding files to and getting files from an IPFS node, we will need some files to experiment with. For this purpose, there is an upload file step in each code exercise of this tutorial.
-
-## Add a file
-
-First let's learn how to add a file to your IPFS node. We'll do this by executing the following expression:
-
-```javascript
-await ipfs.add(data, [options], [callback])
-```
-
-You can find the full documentation for this method [here](https://github.com/ipfs/interface-js-ipfs-core/blob/master/SPEC/FILES.md#add).
-
-It's important to note that whether you want to add a `File` or multiple `File`s, you can pass either a `File` or an array of `File`s as an argument to the `add` method.
-
-The `add` function returns a `Promise`, so you can place an `await` before the function call to block the execution until the return value of the promise is ready to be used.
-
-The result of this `Promise` is an array of objects with the following structure:
-```javascript
-{
-    path: string,
-    hash: string,
-    size: number
-}
-```
-
-This `hash` is what we call a `CID`, a generated address based on the content of the node. For a more in depth look of how `CID`s are generated and what they are, feel free to take a look at the [Decentralized data structures](https://proto.school/#/data-structures) tutorial.
-
-In a future lesson, we will learn how to use this `hash`, or `CID`, to get the contents of a file.
-
+To practice, let's upload one or more files from your computer and take a look at what's been received by the browser as the `files` array.
