+++ src/tutorials/0005/03-exercise.md
@@ -1,9 +1,4 @@
-In this exercise, we will try to read the contents of a file we have already uploaded for you. In order to read its contents, you'll need the path to the file on IPFS, which corresponds to the following `CID`:
+Let's try to add a file to your IPFS node! First, try to upload one or more files.
 
-```
-QmWCscor6qWPdx53zEQmZvQvuWQYxx1ARRCXwYVE4s9wzJ
-```
+**Hint:** You can pass either a `File` or an array of `File`s to the `add` method.
 
-Use the `cat` function to read the contents of the file to a variable.
-
-**Hint:** Don't forget to convert it to a string using `.toString('utf8')` before returning the result.
