+++ src/tutorials/0005/03.md
@@ -1,22 +1,41 @@
-Now that we've learned how to put some files onto your IPFS node, we can use their addresses ([CID's](https://proto.school/#/data-structures/04)) to get their content. If you're familiar with UNIX style commands, you probably know `cat`. The IPFS File API also provides a `cat` function for the same purpose: to get the contents of a file.
 
-## Using `cat`
+## Working with files in ProtoSchool
+Here in our ProtoSchool tutorials, we create a new IPFS node for you in the browser each time you hit the "Submit" button in a lesson. Whenever you see `ipfs.someMethod()` in our lessons, `ipfs` is a variable that refers to your IPFS instance, also known as a node. The actions that you take only affect your own IPFS node, not nodes belonging to your peers.
 
-To get the contents of a file in IPFS, we call the `cat` function:
+We create this special IPFS node behind the scenes, so you can focus on the content of our lessons. To host a fully functional IPFS node of your own, though, IPFS must be installed on your machine, and a local daemon running in your terminal. If you haven't already experimented with this, visit our docs to learn how to [install IPFS](https://docs.ipfs.io/guides/guides/install/) and [initialize your node](https://docs.ipfs.io/introduction/usage/#initialize-the-repository).
 
-```javascript
-await ipfs.cat(ipfsPath, [options], [callback])
-```
+As mentioned previously, the methods discussed in this tutorial are part of the [Files API](https://github.com/ipfs/interface-js-ipfs-core/blob/master/SPEC/FILES.md). Check the documentation for more specific details, such as options for each API function.
 
-The IPFS path can take several formats (you can read about them in the [File API](https://github.com/ipfs/interface-js-ipfs-core/blob/master/SPEC/FILES.md#cat) documentation). In this lesson, we will use the simplest one: the `hash` string present in the objects we receive when the `ipfs.add` function is called.
+## Add a file
 
-The returned data comes in the form of an `Buffer`. If you uploaded a plain text file (such as a `.txt`), you can read the contents of the file by calling `result.toString('utf8')`.
-
-So if you have the `CID` to a text file whose contents you want to read, you can do this:
+First let's learn how to add a file to your IPFS node. We'll do this by executing the following expression:
 
 ```javascript
-let bufferedContents = await ipfs.cat(cid) // returns a buffer
-let contents = bufferedContents.toString('utf8') // a string
+await ipfs.add(data, [options], [callback])
 ```
 
-When you're ready to try this in the real world, you should note that the `cat` method can result in heavy memory usage, depending on the contents of the file being read. If you find this to be the case, you might want to explore the [catReadableStream](https://github.com/ipfs/interface-js-ipfs-core/blob/master/SPEC/FILES.md#catreadablestream) or [catPullStream](https://github.com/ipfs/interface-js-ipfs-core/blob/master/SPEC/FILES.md#catpullstream) methods instead.
+So if we had a file object in our browser, accessible via a variable catPic, and we wanted to add it to our IPFS node, we could do it like so:
+
+```javascript
+await ipfs.add(catPic)
+```
+
+You can find the full documentation for this method [here](https://github.com/ipfs/interface-js-ipfs-core/blob/master/SPEC/FILES.md#add).
+
+It's important to note that whether you want to add a `File` or multiple `File`s, you can pass either a `File` or an array of `File`s as an argument to the `add` method.
+
+The `add` function returns a `Promise`, so you can place an `await` before the function call to block the execution until the return value of the promise is ready to be used.
+
+The result of this `Promise` is an array of objects with the following structure:
+```javascript
+{
+    path: string,
+    hash: string,
+    size: number
+}
+```
+
+This `hash` is what we call a `CID`, a generated address based on the content of the node. For a more in depth look of how `CID`s are generated and what they are, feel free to take a look at the [Decentralized data structures](https://proto.school/#/data-structures) tutorial.
+
+In a future lesson, we will learn how to use this `hash`, or `CID`, to get the contents of a file.
+
