+++ src/tutorials/0005/04.md
@@ -1,27 +1,37 @@
-When adding one or multiple files to the IPFS node, you usually get back one `CID` per file, which you can use to access their contents.
+Now that we've learned how to put some files onto your IPFS node, we can use their addresses ([CID's](https://proto.school/#/data-structures/04)) to get their content. If you're familiar with UNIX style commands, you probably know `cat`. The IPFS File API also provides a `cat` function for the same purpose: to get the contents of a file.
 
-If the files are uploaded into a directory, though, IPFS will generate more than one `CID`. You will receive a path for each file you add, as well as a `CID` for the directory itself. We do this by using a special option when calling the `add` function:
 
-```javascript
-await ipfs.add(file, { wrapWithDirectory: true })
-```
+## How to find a file in an IPFS node
 
-Organizing files into a directory in this way will enable us to do interesting things with the `ls` and `get` methods, which we will talk about in the next two lessons.
-
-The `add` function has many other features, too, which you can check out in the [`add` section](https://github.com/ipfs/interface-js-ipfs-core/blob/master/SPEC/FILES.md#add) of the IPFS File API documentation.
-
-## The file argument
-
-To upload multiple nested directories, the `add` function requires a little more help. We need to provide a path in order for the ipfs node to properly replicate the directory structure.
-
-In these cases, we replace the `file` argument in the `add` function call with an object like:
+As we saw in the previous lesson, the `add` method returned an object -- or an array of objects -- with the following structure:
 
 ```javascript
 {
-    path: '/some/file.txt',
-    content: file
+    path: string,
+    hash: string,
+    size: number
 }
 ```
 
-If we want to add multiple files at once — into a single or multiple directories — we just need to pass an array of objects like the one shown above to the `add` method.
+The `hash` property in this object is a `CID` pointing to the file, which means we can use it to find the file in the IPFS node when we want to read its contents.
 
+## Using `cat`
+
+To get the contents of a file in IPFS, we call the `cat` function:
+
+```javascript
+await ipfs.cat(ipfsPath, [options], [callback])
+```
+
+The IPFS path can take several formats (you can read about them in the [File API](https://github.com/ipfs/interface-js-ipfs-core/blob/master/SPEC/FILES.md#cat) documentation). In this lesson, we will use the simplest one: the `hash` string present in the objects we receive when the `ipfs.add` function is called.
+
+The returned data comes in the form of an `Buffer`. If you uploaded a plain text file (such as a `.txt`), you can read the contents of the file by calling `result.toString('utf8')`.
+
+So if you have the `CID` to a text file whose contents you want to read, you can do this:
+
+```javascript
+let bufferedContents = await ipfs.cat(cid) // returns a buffer
+let contents = bufferedContents.toString('utf8') // a string
+```
+
+When you're ready to try this in the real world, you should note that the `cat` method can result in heavy memory usage, depending on the contents of the file being read. If you find this to be the case, you might want to explore the [catReadableStream](https://github.com/ipfs/interface-js-ipfs-core/blob/master/SPEC/FILES.md#catreadablestream) or [catPullStream](https://github.com/ipfs/interface-js-ipfs-core/blob/master/SPEC/FILES.md#catpullstream) methods instead.
