+++ src/tutorials/0005/05-exercise.md
@@ -1,3 +1,7 @@
-In the previous exercise, we uploaded multiple files into a `/dir` directory. This time, let's check which files are, in fact, in the `/dir` directory of our ipfs node.
+Let's upload some files and add them to our IPFS node. This time, however, we will use the option `{ wrapWithDirectory: true }` to put all the files in the same directory!
 
-**Hint:** The ipfs path (`CID`) you need to call the `ls` function is stored in the `pathCID` variable.
+We want to add the files into a directory called `/dir`. You can upload one or multiple files.
+
+Don't forget, you can get the name of the file you uploaded with `file.name`.
+
+**Hint:** Remember the `add` function from our previous lesson? It looks like `await ipfs.add(file, options)`. Dont forget you need to pass an array to file if you want to add multiple files, and the object structure to indicate the desired path of the file.
