+++ src/tutorials/0005/05.md
@@ -1,10 +1,31 @@
-Now that we've uploaded some files into a directory in our IPFS node, let's learn how we can check the contents of a directory. If you use the terminal frequently, you're familiar with the `ls` command. Well, the Files API provides you with an `ls` function to achieve the same purpose.
+When adding one or multiple files to the IPFS node, you usually get back one `CID` per file, which you can use to access their contents.
 
-Calling `ls` looks like this:
+If the files are uploaded into a directory, though, IPFS will generate more than one `CID`. You will receive a path for each file you add, as well as a `CID` for the directory itself. We do this by using a special option when calling the `add` function:
 
 ```javascript
-await ifps.ls( ipfsPath )
+await ipfs.add(file, { wrapWithDirectory: true })
 ```
 
-Where path can take multiple formats, the simplest being a pure `CID`. More information on the formats of the ipfs path can be found in the [`ls` section](https://github.com/ipfs/interface-js-ipfs-core/blob/master/SPEC/FILES.md#ls)  of the Files API.
+Organizing files into a directory in this way will enable us to do interesting things with the `ls` and `get` methods, which we will talk about in the next two lessons.
+
+The `add` function has many other features, too, which you can check out in the [`add` section](https://github.com/ipfs/interface-js-ipfs-core/blob/master/SPEC/FILES.md#add) of the IPFS File API documentation.
+
+## The file argument
+
+To upload multiple nested directories, the `add` function requires a little more help. We need to provide a path in order for the ipfs node to properly replicate the directory structure.
+
+In these cases, we replace the `file` argument in the `add` function call with an object like:
+
+```javascript
+{
+    path: '/some/file.txt',
+    content: file
+}
+```
+
+If we want to add multiple files at once — into a single or multiple directories — we just need to pass an array of objects like the one shown above to the `add` method.
+
+
+
+The `add` method will return information about all the files you added, as well as details about any directory it had to create in order to correctly add the files in the correct paths.
