+++ src/tutorials/0005/06.md
@@ -1,35 +1,10 @@
-Sometimes we want to know more information about an IPFS path, besides its content (assuming it's a file).
+Now that we've uploaded some files into a directory in our IPFS node, let's learn how we can check the contents of a directory. If you use the terminal frequently, you're familiar with the `ls` command. Well, the Files API provides you with an `ls` function to achieve the same purpose.
 
-For this purpose, we have the `get` method, which we can use like this:
-
-```javascript
-await ipfs.get(ipfsPath)
-```
-
-The result from calling this method is an array of objects with the following structure:
+Calling `ls` looks like this:
 
 ```javascript
-{
-    hash: string,
-    name: string,
-    path: string,
-    depth: number,
-    size: number,
-    type: string, //can be "file" or "dir"
-    content: Buffer, // only exists if the type === "file"
-}
+await ifps.ls( ipfsPath )
 ```
 
-Why is the returned value an array? This is one of the advantages the `get` method has over `cat`. Now that we know how to put files into diferent directories in our IPFS node, it would be interesting to have a way to recursively get the contents of a directory. If we provide an IPFS directory path to the `get` method, the returned array will contain every subdirectory and every file present inside the specified directory.
+Where path can take multiple formats, the simplest being a pure `CID`. More information on the formats of the ipfs path can be found in the [`ls` section](https://github.com/ipfs/interface-js-ipfs-core/blob/master/SPEC/FILES.md#ls)  of the Files API.
 
-## When should I use `get` and when should I use `cat`
-
-The use case for `cat` is simply to get the contents of a file for which we have an IPFS path, while the `get` method provides us with a lot of additional information.
-
- `get` can:
-
-- Recursively retrieve every file and subdirectory in a directory
-- Tell us which entries are directories and which are files
-- Inform us of the size of a file
-- Pinpoint the depth of a file in the directory structure
-- and offer many other details
